split + fix
