+++ mapping.md
@@ -76,7 +76,8 @@ documentation when migrating._
 | `splitAt`          | `-`                | `splitAt`           |
 | `splitWhen`        | `-`                | `splitWhen`         |
 | `sumBy`            | `sumBy`            | `-`                 |
-| `swap`             | `-`                | `swap`              |
+| `swapIndices`      | `-`                | `swap`              |
+| `swapProps`        | `-`                | `swap`              |
 | `take`             | `take`             | `take`              |
 | `takeWhile`        | `takeWhile`        | `takeWhile`         |
 | `times`            | `times`            | `times`             |
