+++ src/index.ts
@@ -87,7 +87,8 @@ export * from './splitAt';
 export * from './splitWhen';
 export * from './stringToPath';
 export * from './sumBy';
-export * from './swap';
+export * from './swapIndices';
+export * from './swapProps';
 export * from './take';
 export * from './takeWhile';
 export * from './times';
